+++ CHANGELOG.md
@@ -6,15 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0]
+
+### Added
+- Added a new workflow `feds` to process fire perimeters using local files.
+- Added dependencies in `pyproject.toml` and `environment.yml`.
+
+### Changed
+- Changed name of the plugin to `hyp3-akfire-safe`.
+
 ## [2.0.0]
 
 ### Added
-- Added a new workflow to pull fire perimeters from VEDA catalog.
+- Added a new workflow `pull_perimeter` to pull fire perimeters from VEDA catalog.
 
 ## [1.0.0]
 
 ### Added
-- Added a new function to download Landsat imagery using a location point and start and end dates.
+- Added a new function `gather_landsat` to download Landsat imagery using a location point and start and end dates.
   
 ## [0.1.0]